requirements.txt
pycryptodome==3.19.0
python-magic-bin==0.4.14
